Uploading landing to include particles.js
Still experimenting with how I want the landing to feel. This is
definitely not it, but a step in the right direction. There needs to be
some level of motion to the bland landing, and until I get my hands on
professional-grade video equipment, that idea is far from me. I would
rather use particles than some stock footage of a guy sitting a desk.
Might take up 3D design shortly.